Splits projects_handler in two
